Make update receipt by devices a sync event
Updates don't need to be pushed asynchronously and handled by a worker,
and they don't need to be handled as part of a "transaction" whereby
they are coupled to the outbound message that initiated the event.
Generally, device replies (at least, in the case of keypads) are
simple acknowledgements to the last message, or convey some state from
the device, but do not need to be processed in the context of the
message last sent. If they do, this could be tracked alternately than
the transaction interface.

The assumption is we can do bookkeeping on the outbound message, as
keypads are simply marked as offline and reset in the event of a
communications error. More complex bookkeeping could be performed when
an update message is received if proposed state is stored, on the basis
of the invariant that a serial device can be associated with at most one
serial bus, and that bus will only ever have one active communication
with one active device at a time.